Cleanup and temp views for iOS
