+++ JellyfinPlayer/Views/LiveTVHomeView.swift
@@ -1,0 +1,15 @@
+/* JellyfinPlayer/Swiftfin is subject to the terms of the Mozilla Public
+ * License, v2.0. If a copy of the MPL was not distributed with this
+ * file, you can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * Copyright 2021 Aiden Vigue & Jellyfin Contributors
+ */
+
+import Stinsen
+import SwiftUI
+
+struct LiveTVHomeView: View {
+    var body: some View {
+        Text("Coming Soon")
+    }
+}
