+++ Shared/Extensions/JellyfinAPIExtensions/BaseItemDtoExtensions.swift
@@ -170,7 +170,34 @@ public extension BaseItemDto {
             return "\(String(progminutes))m"
         }
     }
-
+    
+    func getLiveStartTimeString(formatter: DateFormatter) -> String {
+        if let startDate = self.startDate {
+            return formatter.string(from: startDate)
+        }
+        return " "
+    }
+    
+    func getLiveEndTimeString(formatter: DateFormatter) -> String {
+        if let endDate = self.endDate {
+            return formatter.string(from: endDate)
+        }
+        return " "
+    }
+    
+    func getLiveProgressPercentage() -> Double {
+        if let startDate = self.startDate,
+           let endDate = self.endDate {
+            let start = startDate.timeIntervalSinceReferenceDate
+            let end = endDate.timeIntervalSinceReferenceDate
+            let now = Date().timeIntervalSinceReferenceDate
+            let length = end - start
+            let progress = now - start
+            return progress / length
+        }
+        return 0
+    }
+    
     // MARK: ItemType
 
     enum ItemType: String {
